+++ sysadmin_tools/student_auto_feed/readme.txt
@@ -6,9 +6,11 @@ table data backup and recovery example code is also provided.
 
 Requires at least PHP 5.4 with pgsql, iconv, and ssh2 extensions.
 
+Instructions can be found at http://submitty.org/sysadmin/student_auto_feed
+
 THIS SOFTWARE IS PROVIDED AS IS AND HAS NO GUARANTEE THAT IT IS SAFE OR
 COMPATIBLE WITH YOUR UNIVERSITY'S INFORMATION SYSTEMS.  THIS IS ONLY A CODE
-EXAMPLE FOR YOUR UNIVERSITY'S SYSYTEM'S PROGRAMMER TO PROVIDE AN
+EXAMPLE FOR YOUR UNIVERSITY'S SYSTEM'S PROGRAMMER TO PROVIDE AN
 IMPLEMENTATION.  IT MAY REQUIRE SOME ADDITIONAL MODIFICATION TO SAFELY WORK
 WITH YOUR UNIVERSITY'S AND/OR DEPARTMENT'S INFORMATION SYSTEMS.
 
@@ -39,11 +41,8 @@ will need access to the Submitty course databases and the CSV data dump file.
 
 
 accounts.php
-This is a command line script that will auto-create user authentication accounts
-for all Submitty users.  Authentication requires local Linux user accounts,
-which can also work with other campus authentication mechanisms like PAM and
-Kerberos.  Therefore, submitty_student_auto_feed.php will not create
-authentication access for new students upserted into any course database.
+This is a command line script that will auto-create PAM authentication accounts
+for all Submitty users.  THIS IS NOT NEEDED WITH DATABASE AUTHENTICATION.
 
 accounts.php is also intended to be run as a cron job, but the requirements are
 more stringent.
@@ -55,38 +54,5 @@ more stringent.
 * Recommendation: if this script is run every hour by cronjob, professors can
   advise students who are manually added that they "will have access to Submitty
   within an hour."
-
-
-BACKUP/RECOVERY ****************************************************************
-
-********************************************************************************
-  THESE TOOLS WERE WRITTEN FOR AN EARLIER VERSION OF SUBMITTY AND HAVE NOT YET
-       BEEN UPDATED TO BE COMPATIBLE WITH MORE RECENT DATABASE CHANGES.
-                 USE OF THESE TOOLS IS CURRENTLY NOT ADVISED.
-********************************************************************************
-
-submitty_student_auto_feed.php has data validation checks to help preserve the
-integrity of the courses' database users table from a bad feed.  Should a feed
-of bad data manage to get past validation and corrupt any/all users table, the
-following tools may be able to assist with quick recovery.
-
-
-submitty_users_data_backup.php
-Invoked by driver.php and requires config.php.  If used, this has to be run
-BEFORE submitty_student_auto_feed.php.  The number of days of backups needs to
-be defined in config.php (default: 7).  As users data contains student data
-protected by FERPA, an optional AES encryption feature is also provided
-(disabled by default).  IMPORTANT:  This script will generate an encryption key
-(using /dev/urandom) when a key is not found.  It is vital that this key is
-given very strict access permissions.  If the key is ever leaked, all encrypted
-backups become vulnerable.
-
-
-restore_backup.php
-This utility will revert the users table data of any specific course to a
-backup of a specific date.  This script can also decrypt any encrypted backups.
-Decryption is done entirely in RAM -- there are no temp files made during
-decryption.
-
 
 EOF
